Make required changes in the code
