+++ src/__tests__/Rack.test.js
@@ -1,6 +1,5 @@
-/* eslint-disable*/
 import renderer from 'react-test-renderer';
-import Rack from '../__mocks__/Rack';
+import Rack from '../_mocks_/Rack';
 
 it('renders correctly', () => {
   const tree = renderer
