+++ src/_mocks_/Filter.js
@@ -1,4 +1,3 @@
-/* eslint-disable*/
 import { useState } from 'react';
 import s from './Filter.module.css';
