+++ src/_mocks_/InTheaters.js
@@ -1,4 +1,3 @@
-/* eslint-disable*/
 import React from 'react';
 import { BrowserRouter, Link } from 'react-router-dom';
 import s from './InTheaters.module.css';
