+++ src/_mocks_/store.js
@@ -1,4 +1,3 @@
-/* eslint-disable*/
 const store = {
   topMovies: [
     {
